+++ include/build_version.h
@@ -8,10 +8,10 @@
 #ifndef BUILD_VERSION_H
 #define BUILD_VERSION_H
 
-#define BUILD_NUMBER 1376
-#define BUILD_TIMESTAMP "2026-03-15 22:10:25"
-#define BUILD_DATE "20260315"
-#define GIT_HASH "ddb55e1"
+#define BUILD_NUMBER 1384
+#define BUILD_TIMESTAMP "2026-03-16 22:29:09"
+#define BUILD_DATE "20260316"
+#define GIT_HASH "d0a5c67"
 #define GIT_BRANCH "main"
 
 #endif // BUILD_VERSION_H